Bundle healthcheck with server
