+++ Cargo.toml
@@ -2,7 +2,6 @@
 edition = "2021"
 name = "localauth0"
 version = "0.7.0"
-default-run = "localauth0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
